Use ibrowse instead of httpc as HTTP client
httpc seems to do something odd that Google doesn't like in associate
requests. ibrowse does as well (includes port superflously in the
request, kudos Chandru for spotting the bug) but this is fixed in my
ibrowse fork and in the upcoming ibrowse release.

ibrowse is wrapped by openid_http to hide following redirects and
other minutia.